\name{lassogrp-internal}
\title{Internal Functions of 'lassogrp' Package}
\alias{get.blocks}
\alias{varBlockStand}
\alias{varBlockScale}
\alias{lassogrp.formula}
\alias{lassogrp.default}
\description{
  Internal lassogrp objects, typically \code{\link{function}}s
}
\details{
  These are not to be called by the user, and hence need not be
  understood.
}
\keyword{internal}
